Move shared utility to its own file
